+++ resources/views/includes/modals/group-share-stats.blade.php
@@ -54,20 +54,11 @@
 
                   <div class="form-group">
                       <label for="group_co2_stats_embed">@lang('groups.embed_code_header'):</label>
-                      <input type="text" class="form-control field" id="group_co2_stats_embed" value='<iframe src="{{{ env('APP_URL') }}}/outbound/info/group/{{{ $group->idgroups }}}/manufacture" frameborder="0" width="700" height="850"></iframe>'>
+                      <input type="text" class="form-control field" id="group_co2_stats_embed" value='<iframe src="{{{ env('APP_URL') }}}/outbound/info/group/{{{ $group->idgroups }}}/leaf" frameborder="0" width="700" height="850"></iframe>'>
                   </div>
 
                   <div class="embed-responsive embed-responsive-21by9">
-                    <iframe src="{{{ env('APP_URL') }}}/outbound/info/group/{{{ $group->idgroups }}}/manufacture" frameborder="0" width="700" height="850" class="form-control embed-responsive-item"></iframe>
-                  </div>
-
-                  <div class="form-group">
-                      <label for="group_co2_stats_embed">@lang('groups.embed_code_header'):</label>
-                      <input type="text" class="form-control field" id="group_co2_stats_embed" value='<iframe src="{{{ env('APP_URL') }}}/outbound/info/group/{{{ $group->idgroups }}}/consume" frameborder="0" width="700" height="850"></iframe>'>
-                  </div>
-
-                  <div class="embed-responsive embed-responsive-21by9">
-                    <iframe src="{{{ env('APP_URL') }}}/outbound/info/group/{{{ $group->idgroups }}}/consume" frameborder="0" width="700" height="850" class="form-control embed-responsive-item"></iframe>
+                    <iframe src="{{{ env('APP_URL') }}}/outbound/info/group/{{{ $group->idgroups }}}/leaf" frameborder="0" width="700" height="850" class="form-control embed-responsive-item"></iframe>
                   </div>
 
               </div>
